backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pdfminer.six==20231228
python-docx==0.8.11
PyPDF2==2.10.0
scikit-learn==1.0.2
numpy==1.21.2
pandas==2.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
pydantic==2.6.1
aiofiles==23.2.1
httpx==0.26.0
transformers==4.37.2
torch==2.2.0
spacy==3.5.3
nltk==3.8.1 